models/2026-02-17-pytest/tests/requirements.txt
pytest
pytest-html
matplotlib
ansys-scadeone-core
